Add project creation on VM creation

+++ core/sql/migrations/001_user_project/01_data_center_ddl_init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE data_center (
-    data_center_id	SMALLINT,
+    data_center_id	INT,
   	description		VARCHAR(255),
   	
   	PRIMARY KEY (data_center_id)
